Added Support to DeepStream 7.1/ Added Support Python Apps including YOUTUBE videos
