+++ config/pgie/config_pgie_yolo_det.txt
@@ -3,9 +3,9 @@ gpu-id=0
 net-scale-factor=0.0039215697906911373
 model-color-format=0
 onnx-file=/apps/deepstream-yolo-e2e/models/yolov10n-trt.onnx
-model-engine-file=/apps/deepstream-yolo-e2e/models/yolov10n-trt-fp16-netsize-640-batch-1.engine
+model-engine-file=/apps/deepstream-yolo-e2e/models/yolov10n-trt-fp16-netsize-640-batch-4.engine
 labelfile-path=/apps/deepstream-yolo-e2e/models/labels.txt
-batch-size=1
+batch-size=4
 infer-dims=3;640;640
 force-implicit-batch-dim=0
 # 0: FP32 1: INT8 2: FP16
