+++ config/pgie/config_pgie_yolo_seg.txt
@@ -2,10 +2,10 @@
 gpu-id=0
 net-scale-factor=0.0039215697906911373
 model-color-format=0
-onnx-file=models/yolov9-c-seg-trt.onnx
-model-engine-file=models/yolov9-c-seg-trt-netsize-640-batch-1.engine
-labelfile-path=labels.txt
-batch-size=1
+onnx-file=/apps/deepstream-yolo-e2e/models/yolov9-c-seg-trt.onnx
+model-engine-file=/apps/deepstream-yolo-e2e/models/yolov9-c-seg-trt-fp16-netsize-640-batch-4.engine
+labelfile-path=/apps/deepstream-yolo-e2e/models/labels.txt
+batch-size=4
 force-implicit-batch-dim=0
 infer-dims=3;640;640
 # 0: FP32 1: INT8 2: FP16
@@ -21,7 +21,7 @@ cluster-mode=4
 maintain-aspect-ratio=1
 symmetric-padding=1
 parse-bbox-instance-mask-func-name=NvDsInferYoloMask
-custom-lib-path=nvdsinfer_yolo/libnvds_infer_yolo.so
+custom-lib-path=/apps/deepstream-yolo-e2e/nvdsinfer_yolo/libnvds_infer_yolo.so
 output-instance-mask=1
 segmentation-threshold=0.5
