+++ python_apps/config/config.ini
@@ -1,0 +1,17 @@
+[Settings]
+MUXER_BATCH_TIMEOUT_USEC = 33000
+MUXER_OUTPUT_WIDTH = 1280
+MUXER_OUTPUT_HEIGHT = 720
+
+TILED_OUTPUT_WIDTH = 1280
+TILED_OUTPUT_HEIGHT = 720
+
+OSD_PROCESS_MODE = 0
+OSD_DISPLAY_TEXT = 1
+
+OUTPUT_DIRECTORY = ./videos_output
+OUTPUT_PREFIX = deepstream_out
+
+RTSP_PORT = 8554
+RTSP_FACTORY = /live
+RTSP_UDPSYNC = 8255
