+++ python_apps/config/media.ini
@@ -1,0 +1,34 @@
+[MediaSettings-0]
+type = file  
+url = /opt/nvidia/deepstream/deepstream/samples/streams/sample_1080p_h264.mp4    
+enable = 0   
+
+[MediaSettings-2]
+type = file   
+url = /opt/nvidia/deepstream/deepstream/samples/streams/sample_ride_bike.mov
+enable = 0  
+
+[MediaSettings-3]
+type = rtsp 
+url = rtsp://localhost/live   
+enable = 0   
+
+[MediaSettings-4]
+type = youtube 
+url = https://www.youtube.com/watch?v=cH7VBI4QQzA
+enable = 1 
+
+[MediaSettings-5]
+type = youtube 
+url = https://www.youtube.com/watch?v=KSsfLxP-A9g
+enable = 0 
+
+[MediaSettings-6]
+type = youtube 
+url = https://www.youtube.com/watch?v=yreUvpVjf0Q
+enable = 0 
+
+[MediaSettings-7]
+type = youtube 
+url = https://www.youtube.com/watch?v=2LE3DODbJAQ
+enable = 0
